+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features
 
 * Better rendering of run errors
+* Add `check()` function to ensure requirements are met
+* Add `ProblemException` to handle problems in a more structured way
 
 ### Internal